refactor: extract QuantityStepper from ItemRow
