+++ crates/frontend/src/components/items/mod.rs
@@ -6,3 +6,4 @@ pub mod date_item_row;
 pub mod inline_date_editor_section;
 pub mod item_actions;
 pub mod item_row;
+pub mod quantity_stepper;
